ADLB changes to get ADLB type extra info when initializing a struct.
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/branches/issue-586@11053 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/adlb-defs.h
@@ -74,7 +74,7 @@ typedef enum
 typedef short adlb_data_type_short;
 
 // Identifier for sub-types of ADLB struct
-typedef int adlb_struct_type;
+typedef int16_t adlb_struct_type;
 
 // Additional type info for particular types
 typedef struct {
